Update oplsaa.xml
When trying to apply OPLSAA forcefield to simulation box with water, there were no angle parameters for the water molecules, since in atom types there was only OW and HW from tip4p model, but there was no angle parameters for "HW_tip4pew" and "OW_tip4pew" bond. 
I was able to apply the forcfield adding the tip3p model for water with its according reference and overriding the SMARTS definition for the OH in alcohols since the oxygen has a : [O;X2](H) definition.